+++ Intent.Modules.Java.Services/Templates/Enum/EnumTemplatePartial.cs
@@ -29,7 +29,7 @@ namespace Intent.Modules.Java.Services.Templates.Enum
         public override ITemplateFileConfig GetTemplateFileConfig()
         {
             return new JavaFileConfig(
-                className: $"{Model.Name}",
+                className: $"{Model.Name.ToPascalCase()}",
                 package: this.GetPackage(),
                 relativeLocation: this.GetFolderPath()
             );