+++ Intent.Modules.Java.Domain/Templates/DomainModel/DomainModelTemplate.cs
@@ -83,7 +83,7 @@ namespace Intent.Modules.Java.Domain.Templates.DomainModel
             #line hidden
             
             #line 19 "C:\Dev\Intent.Modules.Java\Intent.Modules.Java.Domain\Templates\DomainModel\DomainModelTemplate.tt"
-            this.Write(this.ToStringHelper.ToStringWithCulture(GetBaseClass()));
+            this.Write(this.ToStringHelper.ToStringWithCulture(GetBaseType()));
             
             #line default
             #line hidden